fix: onedrive uninstaller failing on some users due to non existing folder

+++ src/playbook/Executables/ONED.cmd
@@ -1,13 +1,3 @@
-@echo off
-
-for /f "usebackq delims=" %%a in (`dir /b /a:d "%SystemDrive%\Users"`) do (
-	if exist "%SystemDrive%\Users\%%a\OneDrive" (
-		dir "%SystemDrive%\Users\%%a\OneDrive" /b | findstr "." > nul 2>&1 && (
-			echo Not stripping OneDrive as OneDrive files exist, exiting...
-			exit 6000
-		)
-	)
-)
 
 taskkill /f /im OneDrive.exe > nul 2>&1
 for %%a in (
@@ -29,18 +19,22 @@ for /f "usebackq tokens=2 delims=\" %%a in (`reg query HKU ^| findstr /r /x /c:"
 
 rmdir /q /s "%ProgramData%\Microsoft OneDrive" > nul 2>&1
 rmdir /q /s "%LOCALAPPDATA%\Microsoft\OneDrive" > nul 2>&1
+rmdir /q /s "%SystemDrive%\OneDriveTemp" > nul 2>&1
 
 for /f "usebackq delims=" %%a in (`dir /b /a:d "%SystemDrive%\Users"`) do (
-	rmdir /q /s "%SystemDrive%\Users\%%a\AppData\Local\Microsoft\OneDrive" > nul 2>&1
-	rmdir /q /s "%SystemDrive%\Users\%%a\OneDrive" > nul 2>&1
-	del /q /f "%SystemDrive%\Users\%%a\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\OneDrive.lnk" > nul 2>&1
+	if exist "%SystemDrive%\Users\%%a\AppData\Local\Microsoft\OneDrive" rmdir /q /s "%SystemDrive%\Users\%%a\AppData\Local\Microsoft\OneDrive" > nul 2>&1
+	if exist "%SystemDrive%\Users\%%a\OneDrive" rmdir /q /s "%SystemDrive%\Users\%%a\OneDrive" > nul 2>&1
+	if exist "%SystemDrive%\Users\%%a\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\OneDrive.lnk" del /q /f "%SystemDrive%\Users\%%a\AppData\Roaming\Microsoft\Windows\Start Menu\Programs\OneDrive.lnk" > nul 2>&1
 )
 
-for /f "usebackq delims=" %%a in (`reg query "HKLM\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\SyncRootManager" ^| findstr /i /c:"OneDrive"`) do reg delete "%%a" /f > nul 2>&1
+for /f "usebackq delims=" %%a in (`reg query "HKLM\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\SyncRootManager" 2^>nul ^| findstr /i /c:"OneDrive"`) do reg delete "%%a" /f > nul 2>&1
 
-for /f "tokens=2 delims=\" %%a in ('schtasks /query /fo list /v ^| findstr /c:"\OneDrive Reporting Task" /c:"\OneDrive Standalone Update Task"') do (
+for /f "tokens=2 delims=\" %%a in ('schtasks /query /fo list /v 2^>nul ^| findstr /c:"\OneDrive Reporting Task" /c:"\OneDrive Standalone Update Task"') do (
 	schtasks /delete /tn "%%a" /f > nul 2>&1
 )
+
+reg delete "HKLM\SOFTWARE\Classes\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /f > nul 2>&1
+reg delete "HKLM\SOFTWARE\Classes\WOW6432Node\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /f > nul 2>&1
 
 exit /b
 
@@ -58,8 +52,8 @@ for /f "usebackq delims=" %%a in (`reg query "HKU\%~1\SOFTWARE\Microsoft\Windows
 	reg delete "%%a" /f > nul 2>&1
 )
 
-reg add "HKU\%~1\SOFTWARE\Classes\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /v "System.IsPinnedToNameSpaceTree" /t REG_DWORD /d "0" /f > nul 2>&1
-reg add "HKU\%~1\SOFTWARE\Classes\WOW6432Node\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /v "System.IsPinnedToNameSpaceTree" /t REG_DWORD /d "0" /f > nul 2>&1
+reg delete "HKU\%~1\SOFTWARE\Classes\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /f > nul 2>&1
+reg delete "HKU\%~1\SOFTWARE\Classes\WOW6432Node\CLSID\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /f > nul 2>&1
 reg delete "HKU\%~1\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\Desktop\NameSpace\{018D5C66-4533-4307-9B53-224DE2ED1FE6}" /f > nul 2>&1
 
 reg delete "HKU\%~1\Environment" /v "OneDrive" /f > nul 2>&1
